backend/delivery-services/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
# mysqlclient==2.0.3  # MySQL database driver
python-dotenv==0.19.2  # For environment variable handling
